+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/chapters/ChaptersPage.kt
@@ -1,7 +1,9 @@
 package au.com.shiftyjelly.pocketcasts.player.view.chapters
 
+import android.os.Build
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.asPaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
@@ -33,7 +35,11 @@ fun ChaptersPage(
 ) {
     LazyColumn(
         state = lazyListState,
-        contentPadding = WindowInsets.navigationBars.asPaddingValues(),
+        contentPadding = if (Build.VERSION.SDK_INT > 29) {
+            WindowInsets.navigationBars.asPaddingValues()
+        } else {
+            PaddingValues(bottom = 56.dp)
+        },
         modifier = modifier
             .background(LocalChaptersTheme.current.background)
             .fillMaxSize()